+++ src/containers/Profile/Gallery/Gallery.js
@@ -10,14 +10,13 @@ import {
 } from 'react-native';
 import Icon from 'react-native-vector-icons/AntDesign';
 import ImagePicker from 'react-native-image-crop-picker';
-import NoResult from 'src/components/NoResult';
 import ImageItem from './ImageItem';
 import {Actions} from 'react-native-router-flux';
 import appConfig from 'app-config';
 import ProfileContext from '../ProfileContext';
 
 const MAX_IMAGE_UPLOADED = 9;
-const ORIGIN_PADDING = 15;
+const ORIGIN_PADDING = 5;
 const ORIGIN_HEIGHT =
   appConfig.device.height -
   (StatusBar.currentHeight || 0) -
@@ -40,7 +39,11 @@ const CHOOSE_PHOTO_DATA = [
 ];
 
 class Gallery extends Component {
+  static defaultProps = {
+    renderHeader: () => {},
+  };
   static contextType = ProfileContext;
+
   state = {
     headingHeight: undefined,
     galleryHeight: undefined,
@@ -50,6 +53,9 @@ class Gallery extends Component {
   refModal = React.createRef();
   unmounted = false;
   uploaded = false;
+
+  moreActions = [this.props.t('common:delete'), this.props.t('common:cancel')];
+  destructiveButtonIndex = 0; // index of delete action.
 
   get imagesLength() {
     return this.props.data?.length || 0;
@@ -141,6 +147,27 @@ class Gallery extends Component {
     });
   }
 
+  handlePressAction = (image, actionIndex) => {
+    switch (actionIndex) {
+      case this.destructiveButtonIndex:
+        Actions.pop();
+        this.props.onDeleteImage(image.id);
+        break;
+    }
+  };
+
+  handlePressImage = (image, index) => {
+    Actions.push(appConfig.routes.itemImageViewer, {
+      images: this.props.data,
+      index,
+      moreActionSheetOptions: {
+        options: this.moreActions,
+        destructiveButtonIndex: this.destructiveButtonIndex,
+        onPress: (actionIndex) => this.handlePressAction(image, actionIndex),
+      },
+    });
+  };
+
   renderItem = ({item: img, index}) => {
     const size = (appConfig.device.width - ORIGIN_PADDING * 4) / 3;
     const sizeStyle = {
@@ -160,7 +187,8 @@ class Gallery extends Component {
         style={[sizeStyle, paddingTopStyle]}
         originPadding={ORIGIN_PADDING}
         img={img.name}
-        onDelete={() => this.props.onDeleteImage(img.id)}
+        onPress={() => this.handlePressImage(img, index)}
+        // onDelete={() => this.props.onDeleteImage(img.id)}
       />
     );
   };
@@ -173,42 +201,15 @@ class Gallery extends Component {
     this.setState({galleryHeight: e.nativeEvent.layout.height});
   };
 
-  render() {
+  renderHeader = () => {
     const {isMainUser} = this.context;
-    // const containerStyle = {
-    //   height: this.containerHeight,
-    //   transform: [{ translateY: -this.paddingAnimatedArea }]
-    // };
-    // const animatedStyle = this.state.headingHeight && {
-    //   transform: [
-    //     {
-    //       translateY: this.state.animatedScroll.interpolate({
-    //         inputRange: [0, this.paddingAnimatedArea],
-    //         outputRange: [0, -this.paddingAnimatedArea],
-    //         extrapolate: 'clamp'
-    //       })
-    //     }
-    //   ]
-    // };
-    const emptyContainerStyle = {
-      flex: 1,
-      // marginTop: this.state.headingHeight || 0,
-      // width: '100%',
-      // height: '100%',
-      backgroundColor: '#fafafa',
-      justifyContent: 'center',
-      alignItems: 'center',
-      // paddingTop: this.paddingAnimatedArea,
-    };
 
     return (
       <>
+        {this.props.renderHeader()}
         <Animated.View
           onLayout={this.onHeadingLayout}
-          style={[
-            styles.headerContainer,
-            // animatedStyle
-          ]}>
+          style={[styles.headerContainer]}>
           {this.props.headerComponent}
 
           <View style={[styles.row]}>
@@ -216,50 +217,40 @@ class Gallery extends Component {
             {isMainUser && (
               <TouchableOpacity onPress={this.handleOpenUploadSelection}>
                 <Text style={styles.uploadText}>
-                  <Icon name="upload" /> Tải thêm ảnh
+                  <Icon name="upload" size={13} />
+                  {`  `}Tải thêm ảnh
                 </Text>
               </TouchableOpacity>
             )}
           </View>
         </Animated.View>
-        <View
-          onLayout={this.onGalleryLayout}
-          style={[
-            styles.container,
-            // containerStyle
-          ]}>
-          {this.imagesLength === 0 ? (
-            <View style={emptyContainerStyle}>
-              <Text style={styles.noImage}>
-                Chưa có ảnh
-              </Text>
-            </View>
-          ) : (
-            <View>
-              <Animated.FlatList
-                showsVerticalScrollIndicator={false}
-                data={this.props.data}
-                scrollEventThrottle={16}
-                // onScroll={Animated.event(
-                //   [
-                //     {
-                //       nativeEvent: {
-                //         contentOffset: {
-                //           y: this.state.animatedScroll
-                //         }
-                //       }
-                //     }
-                //   ],
-                //   { useNativeDriver: true }
-                // )}
-                renderItem={this.renderItem}
-                numColumns={3}
-                keyExtractor={(item, index) => index.toString()}
-                columnWrapperStyle={styles.columnWrapperStyle}
-                // contentContainerStyle={{ paddingTop: this.paddingAnimatedArea }}
-              />
-            </View>
-          )}
+      </>
+    );
+  };
+
+  renderEmpty = () => {
+    return (
+      <View style={styles.emptyContainerStyle}>
+        <Text style={styles.noImage}>Chưa có ảnh</Text>
+      </View>
+    );
+  };
+
+  render() {
+    return (
+      <>
+        <View onLayout={this.onGalleryLayout} style={[styles.container]}>
+          <Animated.FlatList
+            showsVerticalScrollIndicator={false}
+            data={this.props.data || []}
+            scrollEventThrottle={16}
+            renderItem={this.renderItem}
+            numColumns={3}
+            keyExtractor={(item, index) => index.toString()}
+            columnWrapperStyle={styles.columnWrapperStyle}
+            ListHeaderComponent={this.renderHeader}
+            {...this.props.listProps}
+          />
         </View>
       </>
     );
@@ -269,23 +260,32 @@ class Gallery extends Component {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    zIndex: 1,
+    backgroundColor: '#fcfcfc',
+  },
+  emptyContainerStyle: {
+    flex: 1,
+    backgroundColor: '#fafafa',
+    justifyContent: 'center',
+    alignItems: 'center',
   },
   headerContainer: {
     zIndex: 2,
   },
   row: {
     backgroundColor: '#f1f1f1',
+    backgroundColor: '#fff',
     flexDirection: 'row',
     justifyContent: 'space-between',
     alignItems: 'flex-end',
     padding: 15,
-    borderBottomColor: '#ccc',
-    borderBottomWidth: 0.5,
+    borderTopColor: appConfig.colors.border,
+    borderTopWidth: 0.5,
   },
   title: {
     fontWeight: '600',
-    color: '#444',
+    color: '#333',
+    fontSize: 16,
+    letterSpacing: 1,
   },
   uploadText: {
     color: '#777',
@@ -301,7 +301,7 @@ const styles = StyleSheet.create({
     fontStyle: 'italic',
     color: '#666',
     fontWeight: '300',
-    textAlign: 'center'
+    textAlign: 'center',
   },
 });